Added GL Quote to twl list view
